UmapitaCustomGroupBox を umapita_custom_group_box.h に分離
